% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{honupsy_weights}
\alias{honupsy_weights}
\title{HoNOS-based Weights for Nursing Hours Estimation}
\format{
A data frame with X rows and Y variables:
\describe{
\item{var1}{Description of variable 1.}
\item{var2}{Description of variable 2.}
}
}
\source{
Author(s) (Year). Title. \emph{Journal}, Volume(Issue), pages.
\doi{10.xxxx/xxxxx}
}
\usage{
honupsy_weights
}
\description{
Item weights derived from a published study on Swiss inpatient psychiatric
data, used to estimate registered nurse (RN) hours based on
HoNOS severity scores.
}
\keyword{datasets}
